requirements.txt
Flask
python-dotenv
scikit-learn
numpy
nltk
regex
chardet
ftfy
wordninja
requests